+++ packages/preview-middleware/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/preview-middleware
 
+## 0.16.103
+
+### Patch Changes
+
+-   Updated dependencies [5a68903]
+    -   @sap-ux/project-access@1.28.4
+    -   @sap-ux/adp-tooling@0.12.69
+
 ## 0.16.102
 
 ### Patch Changes